requirements.txt
aiogram==2.25.1
Flask
SQLAlchemy
psycopg2-binary
cryptography
apscheduler
requests
python-dotenv
gunicorn